STEER/requirements.txt
absl-py==0.15.0
adal==1.2.7
adlfs==2021.7.0
aiohttp==3.7.4.post0
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.6.6
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.7
autovizwidget==0.18.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
black==20.8b1
blackcellmagic==0.0.2
bleach==3.3.0
blinker==1.4
bokeh==2.3.3
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
clang==5.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
ConfigArgParse==1.5.1
cryptography==3.4.7
cycler==0.11.0
Cython==0.29.23
cytoolz==0.11.0
dask==2021.7.0
decorator==4.4.2
defusedxml==0.7.1
distributed==2021.7.0
entrypoints==0.3
fastcluster==1.2.4
findspark==1.4.2
flatbuffers==1.12
fonttools==4.29.1
fsspec==2021.6.1
gast==0.4.0
gensim==3.8.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.1.0
hdijupyterutils==0.18.0
HeapDict==1.0.1
idna==3.3
importlib-metadata==4.11.2
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
isort==5.9.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.3.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
libclang==13.0.0
locket==0.2.0
lxml==4.6.2
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
mistune==0.8.4
mock==4.0.3
msal==1.12.0
msal-extensions==0.3.0
msgpack==1.0.2
msrest==0.6.21
multidict==5.1.0
munkres==1.1.4
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.6.2
nose==1.3.7
notebook==6.2.0
numpy==1.22.2
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.0
pandocfilters==1.4.2
parso==0.8.1
partd==1.2.0
pathspec==0.8.1
pickleshare==0.7.5
Pillow==9.0.1
pip==21.0.1
plotly==4.14.3
portalocker==1.7.0
prometheus-client==0.9.0
prompt-toolkit==3.0.17
protobuf==3.19.4
psutil==5.8.0
py4j==0.10.9
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.1.0
pylint==2.9.6
pyOpenSSL==20.0.1
pyparsing==3.0.7
pyrsistent==0.17.3
PySocks==1.7.1
pyspark==3.1.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-Levenshtein==0.12.2
pytz==2021.1
#pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.7.1
requests==2.27.1
requests-kerberos==0.12.0
requests-oauthlib==1.3.1
retrying==1.3.3
rsa==4.8
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.2
Send2Trash==1.5.0
#setuptools==49.6.0.post20210108
six==1.15.0
smart-open==5.1.0
snorkel==0.9.8
sortedcontainers==2.4.0
sparkmagic==0.13.1
sql-metadata==2.3.0
sqlparse==0.4.2
tabulate==0.8.9
tblib==1.7.0
tdda==1.0.32
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.2
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
#torch==1.4.0
#torchvision==0.5.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
winkerberos==0.7.0
wrapt==1.12.1
yapf==0.31.0
yarl==1.6.3
zict==2.0.0
zipp==3.7.0